# asset_openapi.yaml
openapi: "3.2.0"
info:
  title: AssetPilot API
  version: "4.110.1"
  description: |
    個人資產管理平台 AssetPilot 的完整 API Schema。
    供 Cloudflare API Shield Schema Validation 使用。
    認證方式：JWT 儲存於 httpOnly Cookie（token），除 /api/auth/* 與 /api/config 外均需認證。
    記帳 API 可用 X-Ledger-Id header 或 ledgerId query 選擇帳本，未指定則使用個人帳本。
    每次請求皆驗證登入者的帳本成員資格；viewer 僅可 GET/HEAD/OPTIONS。
    個人偏好、投資、登入憑證及 MCP/API Token/LINE 整合不採用共享帳本選擇。
servers:
  - url: https://asset.shao.one
    description: Production

# ─────────────────────────────────────────
# 標籤定義（對應 Cloudflare API Shield Endpoint Labels）
# ─────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
    LedgerId:
      in: path
      name: ledgerId
      required: true
      schema:
        type: string
...
  schemas:
    Ledger:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        isShared: { type: boolean }
        isPersonal: { type: boolean }
        ownerUserId: { type: string }
        role: { type: string, enum: [owner, editor, viewer] }
        memberCount: { type: integer }
        createdAt: { type: integer }
...
    LedgerMemberTarget:
      type: object
      required: [userId]
      properties:
        userId: { type: string }

    # ── 通用 ──
...
    CategoryInput:
      type: object
      required: [name, type, color]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [income, expense]
        color:
          type: string
        parentId:
          type: string
        sortOrder:
          type: integer

    # ── 帳戶 ──
...
    TransactionInput:
      type: object
      required: [type, amount, date, categoryId, accountId]
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        originalAmount:
          type: number
        currency:
          type: string
        fxRate:
          type: number
        date:
          type: string
        categoryId:
          type: string
        accountId:
          type: string
        note:
          type: string
        excludeFromStats:
          type: boolean
        fxFee:
          type: number
          description: 海外刷卡手續費（TWD）
        clientRef:
          type: string
          pattern: "^[a-f0-9]{32}$"
          description: |
            離線記帳的 idempotency key（007-pwa-offline-entry）。離線新增的交易由
            client 產生此值；伺服器以 (user_id, client_ref) 唯一索引去重，同一
            clientRef 重送不會產生重複交易，並回傳與新建相同的結果。線上直接新增時省略；
            格式不符時視為未提供。

    # ── 預算 ──
...
security:
  - cookieAuth: []

# ─────────────────────────────────────────
# 路由定義
# ─────────────────────────────────────────
paths:
  /api/ledgers:
    get:
      tags: [Ledgers]
      summary: 列出目前登入者加入的帳本
      responses:
        "200":
          description: 個人與共享帳本清單（不包含其他使用者的帳本）
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Ledger" }
    post:
      tags: [Ledgers]
      summary: 建立空白共享帳本，不複製個人資料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
      responses:
        "201": { description: 已建立帳本，建立者為 owner }
        "400": { description: 帳本名稱無效 }
...
  /api/ledgers/{ledgerId}/transfer:
    parameters:
      - { $ref: "#/components/parameters/LedgerId" }
    post:
      tags: [Ledgers]
      summary: owner 移交給另一位有效成員，原 owner 變為 editor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LedgerMemberTarget" }
      responses:
        "200": { description: 已原子移交擁有權，財務列的歸屬不變 }
        "403": { description: 非 owner }
        "404": { description: 目標不是有效帳本成員 }
...
  /api/categories:
    get:
      operationId: getCategories
      tags: [Categories]
      summary: 取得分類清單
      responses:
        "200":
          description: 分類陣列
    post:
      operationId: createCategory
      tags: [Categories]
      summary: 新增分類
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: 新增成功
  /api/categories/{id}:
    put:
      operationId: updateCategory
      tags: [Categories]
      summary: 更新分類
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
                isHidden:
                  type: boolean
                sortOrder:
                  type: integer
      responses:
        "200":
          description: 更新成功
        "404":
          description: 分類不存在
    delete:
      operationId: deleteCategory
      tags: [Categories]
      summary: 刪除分類
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 刪除成功

  # ── 匯率 ──────────────────────────────
...
  /api/transactions:
    get:
      operationId: getTransactions
      tags: [Transactions]
      summary: 取得交易記錄（支援篩選分頁）
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: accountId
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: |
            交易清單（含分頁資訊）。每筆交易含唯讀衍生欄位 repaymentSummaryId
            （非還款交易為空字串；非空代表該筆屬於某次信用卡還款摘要，可作為開啟摘要的入口）。
    post:
      operationId: createTransaction
      tags: [Transactions]
      summary: 新增交易
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "200":
          description: 新增成功
...
  /api/transactions/{id}/restore-ai-note:
    get:
      operationId: previewAiNoteRestore
      tags: [Transactions]
      summary: 預覽「備註經 AI 修改」還原後的內容（唯讀）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 預覽內容
        "404":
          description: 找不到此交易，或不屬於目前使用者
    post:
      operationId: restoreAiNote
      tags: [Transactions]
      summary: 確認執行備註還原
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedUpdatedAt]
              properties:
                expectedUpdatedAt:
                  type: integer
      responses:
        "200":
          description: 已還原
        "400":
          description: 缺少或格式錯誤的 expectedUpdatedAt
        "404":
          description: 找不到此交易，或不屬於目前使用者
        "409":
          description: NoteChanged（樂觀鎖衝突）或 NotRestorable（無可還原快照），皆不變更任何資料
...